requirements.txt
python-dotenv
streamlit
pillow
transformers
openai
torch
ultralytics
numpy